OK, worlRegistry -> WorldRegistry
